refactor: global type 분리
